Add basic Visitor pattern implementation
The Visitor pattern has been added. Since this is the first pattern that
cuts across the currently active file, a few changes to the core pattern
functionality had to be made:

1. Patterns can take over the responsibility for starting to record
modifications and writing to file from InsertPattern (which only
modifies the currently active file). They do so by setting a
SrcCodeModMethod instance in PatternConfigData to non-null.

2. An implementation of that interface for 'modifying behaviour' (see
JDT documentation) has been added via SrcCodeModMethodModifying.

3. InsertionData is a lot like the old InsertionHelper (who has been
provisionally renamed InsertionDataDefault, awaiting reworking) except
generalized in that it does not relate to the currently active file.
It's one of the few parts that need more clean up though.

4. I have added a hack-y subclass to customize ContentProposalAdapter.
Not pretty but it works for now.

Note that there preVisit() and postVisit() methods in each Visitor class
for code that cuts across all visitor-accepting classes (which we dub
'visitees'). The preVisit() method can be used to cause skipping of
visit(), if it returns false.

Functionality for switching between inserting visitor and visitee code
in the currently active file awaits implementation.

Lastly, there is some strange behaviour by Eclipse in that it often
fails to indicate unsaved changes in non-active windows. When selected,
this changes but you cannot invoke Undo on the insertion process for
this file.